# Speed up ECJ tasks by avoiding --release