LUCENE-10185 caused a large performance regression in ECJ tasks by using the --release flag. See graph: http://people.apache.org/~mikemccand/lucenebench/antcleantest.html

We recovered some, but not all, of the performance with https://github.com/apache/lucene/pull/483, but it is still 50% slower.

Instead of using --release, we can just disable "terminal deprecation", and leave this check to `javac`. That single check makes this tool run 50% slower, so it isn't worth it.

I've confirmed that the change fully recovers the build performance drop that was lost from LUCENE-10185:
```
Trunk:
Aggregate task times (possibly running in parallel!):
 160.88 sec.  ecjLintTest
 149.60 sec.  ecjLintMain

Patch:
Aggregate task times (possibly running in parallel!):
 112.95 sec.  ecjLintMain
 105.10 sec.  ecjLintTest
```