@rmuir 2021-11-28
note: the change doesn't help any build machines running with java 11, the `--release` flag acts as a no-op in that case and doesn't cause slowdowns. The change only helps ppl running any newer java version (e.g. java 17). But I think this is probably most developers?