@dweiss 2021-11-28
I think ecj folks may be interested in this one - seems crazy that this creates such a difference.