@rmuir 2021-11-28
well, with the latest release, the cost was reduced substantially: https://github.com/apache/lucene/pull/483

so it's trending in the right direction. but still currently 50% slower than if you don't use `--release` (or happen to be running that exact java version)