@uschindler 2021-11-28
We should really go back to release when the echo people have fixed this. So we should leave a Todo.