@rmuir 2021-11-28
> We should really go back to release when the ecj people have fixed this. So we should leave a Todo.

I already added a comment right where we do this about why. It isn't a TODO, just a statement of fact. Currently, there's no way to run with --release without adding slowdowns.